functions/airnow_connector/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
geojson==2.5.0
google-api-core==1.26.3
google-auth==1.28.0
google-cloud-bigquery==2.13.1
google-cloud-core==1.6.0
google-cloud-firestore==2.3.4
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.2
packaging==20.9
pandas==1.2.3
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.4
